Club signatures accounting for the strand break and strand effect

Function for clubbing signatures from the two strands to remove strnad bias. An example is G->A and C->T are clubbed into C->T to remove the bias due to whether 5' or 3' ends of the strands has been sequenced.

club_signature_counts(signature_counts, flanking_bases = 1)

Arguments

signature_counts

The matrix of counts of all signatures as produced by aggregate_bin_counts.

flanking_bases

The number of flanking bases. Defaults to 2.

Value

Returns a matrix of clubbed signatures. The default choice of the mutation signatures are C->T, C->A, C->G, T->A, T->C and T->G. The other sets of mutations like G->A are clubbed to C->T to remove strand bias.