Builds damage logo plots along with mutational profile across read and the strand break composition for the clusters

Damage Logo plots for each cluster representing the substitution frequency of the 6 substitutional patterns (adjusting for strand bias) and the flanking bases arranged as per relative frequency on either side of the substitution along with the probability of the cluster mutational profile along the read and the strand break composition.

damageLogo_five(theta_pool, sig_names = NULL, ic.scale = TRUE,
  max_pos = 20, flanking_bases = 1, yscale_change = TRUE, xaxis = TRUE,
  yaxis = TRUE, xlab = " ", xaxis_fontsize = 5, xlab_fontsize = 10,
  y_fontsize = 10, mut_width = 2, start = 1e-04, renyi_alpha = 1,
  pop_names = paste0("Cluster ", 1:dim(theta_pool)[2]), logoport_x = 0.28,
  logoport_y = 0.5, logoport_width = 0.3, logoport_height = 0.4,
  lineport_x = 0.72, lineport_y = 0.34, lineport_width = 0.25,
  lineport_height = 0.25, breaklogoport_x = 1, breaklogoport_y = 0.4,
  breaklogoport_width = 0.4, breaklogoport_height = 0.5, barport_x = 0.6,
  barport_y = 0.6, barport_width = 0.3, barport_height = 0.25)

Arguments

theta_pool

The theta matrix obtained from running the grade of membership model that stores for each cluster, the probability distribution over all the mutational signatures.

sig_names

The mutational signature names. Defaults to the rownames of the theta matrix above.

ic.scale

A binary variable indicating whether the height of the bars for substitution and flanking bases should be adjusted by the information criterion.

max_pos

The maximum distance from the end of the read upto which mutations are considered.

flanking_bases

The number of flanking bases of the mutational signature.

yscale_change

A binary variable indicating whether the Y axis scale should be adjusted based on the size of the logos, defaults to TRUE.

xaxis

A binary indicating whether the X axis of the logo plot should be shown

yaxis

A binary indicating whether the Y axis of the logo plot should be shown

xlab

The labels for X axis.

xaxis_fontsize

The fontsize of the X axis ticks.

xlab_fontsize

The fontsize of the X axis labels.

y_fontsize

The fontsize of the Y axis ticks.

mut_width

Thw width of the bar for the mutation at the center.

start

The starting point of the stacking of logos on the Y axis. Should be close to 0, defau;ts to 0.0001.

renyi_alpha

The entropy scale for the Renyi entropy on the flanking bases and mutations.

pop_names

The title of the plot. Defaults to the cluster labels.

logoport_x

the X-axis position of the plot window for the logo plot

logoport_y

the Y-axis position of the plot window for the logo plot

logoport_width

the width of the plot window for the logo plot

logoport_height

the width of the plot window for the logo plot

lineport_x

the X-axis position of the plot window for the mutational profile line plot.

lineport_y

the Y-axis position of the plot window for the mutational profile line plot.

lineport_width

the width of the plot window for the mutational profile line plot.

lineport_height

the width of the plot window for the mutational profile line plot.

Value

Returns logo plots for each cluster