Function to perform grid plots of PCA for signature counts with position information from multiple data sources

This fucntion performs PCA on signature counts or proportions data (both normalized or unnormalized) from multiple data soirces, and stored in a list of counts table as obtained from aggregate_bin_counts or club_signature_counts. The data is clubbed into positional bins and scatter plots for PC1 vs PC2, PC2 vs PC3 and PC1 vs PC3 are displayed in a grid.

gridPCA_position_signatures_combo(signature_list, labs, source_names = NULL,
  input_pos = 1:20, normalize = TRUE, cols = c("red", "blue",
  "darkgoldenrod1", "cyan", "firebrick", "green", "hotpink", "burlywood",
  "yellow", "darkgray", "deepskyblue", "darkkhaki", "brown4", "darkorchid",
  "magenta", "yellow", "azure1", "azure4"))

Arguments

signature_list

A list of signature counts or signature proportion matrices from multiple sources of data

labs

The factor labels for the samples used for coloring in the PC plot

source_names

The names of the list of signature counts (equals the number of data sources)

input_pos

The number of positions to be taken into account for building the PCA

normalize

If TRUE, we normalize by the total number of mutations in that sample (analogous to library size normalization in RNA-seq)

cols

The palette used for labeling different data sources in the PCA plot.

Value

Returns grid plot of the PCA plots