Read length distribution plots corresponding to mutationFileFormat files

For each file (of mutationFileFormat type) in a directory, we plot the read length distribution under three settings - all reads, reads containing C to T mutations and reads containing C to T towards the ends of the fragments.

read_length_distribution(dir, end_break = 5, plot_layout = c(3, 3),
  cols = c("red", "green", "blue"), cex_legend = 0.5, cex.main = 0.5)

Arguments

dir

The directory housing all the mutationFileFormat files.

end_break

The position on the read from the end of the fragment such that if C->T change occurs beyond that end break, it is considered in the third class of reads (potentially damaged reads)

plot_layout

The layout of the plot with multiple images (deafult is a 3x3 layout)

cols

The color profiles of the three plots. Defaults to "red", "green" and "blue".

cex_legend

The cex of the legend.

cex_main

The cex of the title of the figure.

Value

The read length distribution plots for each of the mutationFileFormat files in the folder